+++ setup/update/2.0.7.php
@@ -1,0 +1,4 @@
+<?php
+$secret = \Thelia\Tools\TokenProvider::generateToken();
+
+\Thelia\Model\ConfigQuery::write('form.secret', $secret, 0, 0);